+++ _includes/js/simulator.js
@@ -49,6 +49,10 @@ var compute_income = function() {
 
     //calcul de l'indice
     var indice = table_indices[$("#echelon option:selected").val()];
+    //décret 2023-519
+    if(currentMoment >= moment('2024-01-01')) {
+        indice += 5;
+    }
     if(isNaN(indice)){
         indice = 0;
         $("#tooltipEchelon").attr('title', "");
@@ -390,7 +394,7 @@ var compute_income = function() {
 
 //variables par défaut
 var corps = 'ieeac';
-var defaultDate = '01/07/2023';
+var defaultDate = '01/01/2024';
 var protoDate = moment('2017-07-01');
 var transfertDate = moment('2017-01-01');
 var csgDate = moment('2018-01-01');
@@ -555,6 +559,20 @@ var initVar = function() {
         _transfertRetenue = transfertRetenue["2019"];
         _csg_deduc = csg_deduc["2018"];
     } else if (currentDate.localeCompare('01/07/2023') == 0){
+        proto = true;
+        _pcs = pcs["2017"];
+        _activity_rate = activity_rate["2016"];
+        _yearEchelon = "2020";
+        _point_indice = point_indice["2023"];
+        _rsi = rsi["2016"];
+        _prime_tech = prime_tech["2016"];
+        _evs = evs["2023"];
+        _exp = exp["2023"];
+        _partTechIEEAC = partTechIEEAC["2023"];
+        _rpc = rpc_rate["2020"];
+        _transfertRetenue = transfertRetenue["2019"];
+        _csg_deduc = csg_deduc["2018"];
+    } else if (currentDate.localeCompare('01/01/2024') == 0){
         proto = true;
         _pcs = pcs["2017"];
         _activity_rate = activity_rate["2016"];